Fix new markdown syntax errors
Migrate from the old {{}} directives used in Hugo. Create a new Astro
component, `ResponsiveFigure`, for figures, and just use regular syntax
highlighting for the code fences (put the line number starts in a
comment in the code rather than going through all the effort to switch
to Prism, add a prism-plugin for line numbers, and figure out how to
change the line number start). Leave a couple TODOs for things that
would not be quick solutions.