fix date display
Replace `date` with `publishDate` everywhere (not strictly required, but
it's a clearer name and probably idiomatic to Astro).

Add the `dayjs` dependency and re-format the `publishDate` property
at build time so that we don't have to manually take care of the string
formatting. Create a custom `Date` component that writes the
machine-readable date in semantic HTML.