Fix TODO's where possible
Because of
[this](https://github.com/RafidMuhymin/astro-imagetools/issues/106)
issue it wasn't straightforward for me to pass through height/width
picture attributes. I'll revisit those TODO's when it comes time to
work on the styling.

The Yotchare image didn't belong to me and is now unreachable.

I don't want to migrate the jQuery code for the GitHub widget, and I
didn't see any clear replacement component. It's not that important,
so I'll just add a link instead. This is actually better than what we
had, since we are linking to the commit at the time of writing.