Generate RSS feed
Follow the docs
[here](https://docs.astro.build/en/guides/rss/#2-list-of-rss-feed-objects), replacing `slug` with `url` (filed issue:
https://github.com/withastro/docs/issues/1082). I like the more explicit
version of the code because it decouples the naming of frontmatter
fields from the field names used in RSS. The Astro starter template I
used used `publishDate` everywhere, but passing a glob of files to the
`rss()` function requires that the field be named `pubDate`.

Also add the rss file stylesheet as recommended. Pull out the RSS SVG
and reuse it in the site footer.