Portfolio project images
Add images for each project with appropriate alt text.

* Algorithm::AM: https://en.wikipedia.org/wiki/Analogical_modeling
(I'm the original author of the image).
* LEXiTRON Mac: "apple" spelled in Thai, generated from handwriting font
via https://fontmeme.com/thai/; the อ was replaced with the apple logo,
turned around so that the bite takes the place of the little loop on the
right of the letter.
* Kengdic:
https://commons.wikimedia.org/wiki/File:Middle_Class_in_Joseon.jpg
(is listed as being in the public domain, since it's from the 1850's)
* Soar: https://soar.eecs.umich.edu/filedata/fetch?filedataid=53
* Sublime-Soar-Tools: took the photo of Sublime text myself, with Soar
code from [Rosie](
    picture of code from https://github.com/SoarGroup/rosie/
) loaded and the plugin syntax highlighting active
* Jupyter logo is from here:
https://commons.wikimedia.org/wiki/File:Jupyter_logo.svg. Still need to
display proper attribution.
* Japanese segmenter: took a photo of a page from
数学ガール：乱択アルゴリズム page 348, then edited in InkScape.
* stock photos purchased from dreamstime.com, then resized at
https://onlinepngtools.com/fit-png-in-rectangle.

Include the original SVG's made with InkScape. Convert all SVG's to
PNG/JPG with ImageMagick, as I couldn't get the styling to work right
with a mix of astro-imagetools' `Picture` component and plain `<img>`
tags:

    convert -density 1200 -resize 400x400 Jupyter_logo.svg Jupyter_logo.png

For the AM pointer network image, ImageMagick wouldn't output the arrows
for some reason, so I had to export the PNG from InkScape instead.