Don't use Astro's image for the social preview.
The portfolio image is pretty small and probably not the right dimension
for an ideal social preview image, but I don't have a logo yet
(see #3), so this is the best we can do for now.

twitter requires an image with aspect ratio 2:1, so generate a separate
image for use there.