WIP: language selection on homepage
I didn't realize before that I could name a directory `[lang]` and Astro
would generate the paths accordingly, so now I've reorganized to follow
the structure suggested
[here](https://www.defined.net/blog/internationalizing-docs-pages-with-astro/)
and also used in the Astro "docs" example project.

Also taken from the Astro "docs" example project is a new language
selector component. We show it in the header when the "multilingual"
flag is set, and when the user clicks on a language we navigate to the
same page but in the selected language.