strings is not a page
EVERY file processed by Astro that lives under `pages/` is a page, and
because `strings.ts` was also under `[lang]`, Astro expected a
`getStaticPaths` function to be exported from it, and subsequently threw
an error.

While we're here, run formatting on index.astro, and set VSCode to
always format on save.