redirect previous home-en to /en
I was trying to view my own homepage but was always being redirected to
home-en, which is now a 404! So for now we'll redirect to the correct
home from there. Not sure how many months it takes for a browser to
forget a redirect, but we'll leave this here a while.